+++ azurik_mod/xbe_tools/re_recorder.py
@@ -56,7 +56,6 @@ from typing import Any, Iterator
 
 from .ghidra_client import (
     GhidraClient,
-    GhidraClientError,
     GhidraFunction,
     GhidraLabel,
     GhidraProgramInfo,